Move queue processor into separate file
