+++ internal/reconstitution/queueprocessor.go
@@ -1,0 +1,64 @@
+package reconstitution
+
+import (
+	"context"
+
+	"github.com/go-logr/logr"
+	"k8s.io/client-go/util/workqueue"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+type queueProcessor struct {
+	Client  client.Client
+	Queue   workqueue.RateLimitingInterface
+	Recon   *reconstituter
+	Handler Reconciler
+	Logger  logr.Logger
+}
+
+func (q *queueProcessor) Start(ctx context.Context) error {
+	go func() {
+		<-ctx.Done()
+		q.Queue.ShutDown()
+	}()
+	for q.processQueueItem(ctx) {
+	}
+	return nil
+}
+
+func (q *queueProcessor) processQueueItem(ctx context.Context) bool {
+	item, shutdown := q.Queue.Get()
+	if shutdown {
+		return false
+	}
+	defer q.Queue.Done(item)
+
+	req, ok := item.(*Request)
+	if !ok {
+		q.Logger.Error(nil, "failed type assertion in queue processor")
+		return false
+	}
+
+	logger := q.Logger.WithValues("compositionName", req.Composition.Name, "compositionNamespace", req.Composition.Namespace, "resourceKind", req.Resource.Kind, "resourceName", req.Resource.Name, "resourceNamespace", req.Resource.Namespace)
+	ctx = logr.NewContext(ctx, logger)
+
+	result, err := q.Handler.Reconcile(ctx, req)
+	if err != nil {
+		q.Queue.AddRateLimited(item)
+		logger.Error(err, "error while processing queue item")
+		return true
+	}
+	if result.Requeue {
+		q.Queue.Forget(item) // TODO: Maybe omit after first retry to avoid getting stuck in a patch tightloop?
+		q.Queue.Add(item)
+		return true
+	}
+	if result.RequeueAfter != 0 {
+		q.Queue.Forget(item)
+		q.Queue.AddAfter(item, result.RequeueAfter)
+		return true
+	}
+
+	q.Queue.Forget(item)
+	return true
+}
